requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
requests==2.32.3
geopy==2.4.1
python-telegram-bot==20.7
reportlab==4.0.7
python-docx==1.1.2
Pillow==10.2.0